% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{blockr.echarts-package}
\alias{blockr.echarts}
\alias{blockr.echarts-package}
\title{blockr.echarts: Apache ECharts Visualization Blocks for Blockr}
\description{
Extends 'blockr.core' with interactive blocks for data visualization using Apache ECharts via 'echarts4r'. Includes universal chart blocks (scatter, line, bar, area, pie), specialized blocks (heatmap, radar, sankey, gauge, candlestick, calendar, treemap), and a hierarchical filter block (sunburst, treemap, tree).
}
\author{
\strong{Maintainer}: Christoph Sax \email{christoph@cynkra.com} (\href{https://orcid.org/0000-0002-7192-7044}{ORCID})

Authors:
\itemize{
  \item Nicolas Bennett \email{nicolas@cynkra.com}
  \item David Granjon \email{david@cynkra.com}
  \item Mike Page \email{mike@cynkra.com}
}

Other contributors:
\itemize{
  \item  Bristol Myers Squibb [funder]
}

}
\keyword{internal}
