% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{register_echarts_blocks}
\alias{register_echarts_blocks}
\title{Register ECharts blocks}
\usage{
register_echarts_blocks()
}
\value{
NULL (invisibly)
}
\description{
Registers all blocks from the blockr.echarts package.
}
